// public/reactapp/model/Item.ts
export class Item {

    Title?:string;
    Url?: string;
    Gsm?: string;
    Price?: number;
    AllowBack?: boolean;
    StatusId?: boolean;
    IsTrans?: boolean;
    IsConfirmed?: boolean;
    IsFavorite?: boolean;
    CallCount?: number;
    LastCallDate?: Date;
    CreateDate?: Date;
    Location?: string;
    _id?: AAGUID;
    Files?:any;

}